requirements.txt
pkg-resources==0.0.0
psycopg2==2.7.6.1
pyserial==3.4
